Move server start to server.ts
